Split the weather reader.
