+++ lib/weather_info/weather_info.cpp
@@ -2,11 +2,13 @@
 
 namespace Weather
 {
+//---------------------------------------------------------------------------------------------------------------------
 void WeatherInfo::clear()
 {
     name = weather = weather_desc = temperature_c = humidity_perc = counter = "";
 }
 
+//---------------------------------------------------------------------------------------------------------------------
 String WeatherInfo::get_string()
 {
     String n = name + ", " + weather + ": " + weather_desc + ". T: " + temperature_c + " C, H: " + humidity_perc +
