+++ lib/weather_info/weather_info.hpp
@@ -1,3 +1,7 @@
+//---------------------------------------------------------------------------------------------------------------------
+// weather_info.hpp
+// Weather information to be displayed.
+//---------------------------------------------------------------------------------------------------------------------
 #ifndef WEATHER_INFO_HPP
 #define WEATHER_INFO_HPP
